fix bug with unexpected BubbleChart results when using tibbles
due to the way indexing was performed of a single column in the BubbleChart
function, the code would run in an unexpected way and produce plots that are
incorrect when the dat_list passed in was a list of tibbles instead of a list of
data frames. This commit fixes the issues so that tibbles or data frames will
produce the same plotting results.